Agrego servicios, controladores y endpoints probados

+++ servicio-deposito/src/main/java/com/backend/tpi_backend/servicio_deposito/repositories/UbicacionRepository.java
@@ -2,5 +2,5 @@ package com.backend.tpi_backend.servicio_deposito.repositories;
 
 import com.backend.tpi_backend.servicio_deposito.model.Ubicacion;
 
-public interface UbicacionRepository extends BaseRepository<Ubicacion, Integer> {
+public interface UbicacionRepository extends BaseRepository<Ubicacion, Long> {
 }
